+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 * CLEANUP: add station change entries into the database for language and favicon updates
+* DB: use transactions for migrations
 
 ### Fixed
 * CONFIG: replace config in shared memory